booking_api_fastapi/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
